Create `utils.go`
