+++ test/e2e/case1_framework_deployment_test.go
@@ -3,17 +3,9 @@
 package e2e
 
 import (
-	"context"
-	"fmt"
-	"os/exec"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/client-go/dynamic"
 )
 
 const (
@@ -103,93 +95,3 @@ var _ = Describe("Test framework deployment", func() {
 		Expect(deploy).To(BeNil())
 	})
 })
-
-// Kubectl executes kubectl commands
-func Kubectl(args ...string) {
-	cmd := exec.Command("kubectl", args...)
-
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		// in case of failure, print command output (including error)
-		//nolint:forbidigo
-		fmt.Printf("%s\n", output)
-		Fail(fmt.Sprintf("Error: %v", err))
-	}
-}
-
-// GetWithTimeout keeps polling to get the object for timeout seconds until wantFound is met
-// (true for found, false for not found)
-func GetWithTimeout(
-	client dynamic.Interface,
-	gvr schema.GroupVersionResource,
-	name, namespace string,
-	wantFound bool,
-	timeout int,
-) *unstructured.Unstructured {
-	if timeout < 1 {
-		timeout = 1
-	}
-	var obj *unstructured.Unstructured
-
-	Eventually(func() error {
-		var err error
-		namespace := client.Resource(gvr).Namespace(namespace)
-		obj, err = namespace.Get(context.TODO(), name, metav1.GetOptions{})
-		if wantFound && err != nil {
-			return err
-		}
-		if !wantFound && err == nil {
-			return fmt.Errorf("expected to return IsNotFound error")
-		}
-		if !wantFound && err != nil && !errors.IsNotFound(err) {
-			return err
-		}
-
-		return nil
-	}, timeout, 1).Should(BeNil())
-
-	if wantFound {
-		return obj
-	}
-
-	return nil
-}
-
-// ListWithTimeoutByNamespace keeps polling to list the object for timeout seconds until wantFound is met
-// (true for found, false for not found)
-func ListWithTimeoutByNamespace(
-	clientHubDynamic dynamic.Interface,
-	gvr schema.GroupVersionResource,
-	opts metav1.ListOptions,
-	ns string,
-	size int,
-	wantFound bool,
-	timeout int,
-) *unstructured.UnstructuredList {
-	if timeout < 1 {
-		timeout = 1
-	}
-
-	var list *unstructured.UnstructuredList
-
-	Eventually(func() error {
-		var err error
-		list, err = clientHubDynamic.Resource(gvr).Namespace(ns).List(context.TODO(), opts)
-
-		if err != nil {
-			return err
-		}
-
-		if len(list.Items) != size {
-			return fmt.Errorf("list size doesn't match, expected %d actual %d", size, len(list.Items))
-		}
-
-		return nil
-	}, timeout, 1).Should(BeNil())
-
-	if wantFound {
-		return list
-	}
-
-	return nil
-}
